backend/requirements.txt
flask==2.2.3
flask-cors==3.0.10
werkzeug==2.2.3
pymongo==4.3.3
python-dotenv==1.0.0
requests==2.28.2
feedparser==6.0.10
arxiv==1.4.7
apscheduler==3.10.1
feedgen==0.9.0
gunicorn==20.1.0
